feat(webrtc): enhance data channel handling and introduce new session routing features
- Added new `channel_labels` and `session_routing` libraries for improved session management and routing logic.
- Introduced `SendFramed` function for efficient data channel message sending with chunking support.
- Updated `README.md` to reflect new files and functionalities, including channel labels and session routing.
- Refactored `WebRTCManager` to utilize new session routing logic and improved data channel handling.
- Enhanced `LiveVideoProcessor` to streamline filter state resolution and resource management.
- Implemented `accelerator_label` for human-readable accelerator type representation.
- Added `filter_parameter_mapping` for converting filter parameter types and lifting validation rules.
