+++ src/cpp_accelerator/adapters/webrtc/protocol/BUILD
@@ -12,24 +12,12 @@ cc_library(
 )
 
 cc_library(
-    name = "message_codec",
-    srcs = ["message_codec.cpp"],
-    hdrs = ["message_codec.h"],
+    name = "data_channel_envelope",
+    srcs = ["data_channel_envelope.cpp"],
+    hdrs = ["data_channel_envelope.h"],
     visibility = ["//visibility:public"],
     deps = [
         "//proto:image_processor_service_cc_proto",
-        "//src/cpp_accelerator/adapters/webrtc:data_channel_framing",
-        "@com_github_paullouisageneau_libdatachannel//:libdatachannel",
         "@spdlog",
-    ],
-)
-
-cc_library(
-    name = "session_routing",
-    srcs = ["session_routing.cpp"],
-    hdrs = ["session_routing.h"],
-    visibility = ["//visibility:public"],
-    deps = [
-        "//proto:image_processor_service_cc_proto",
     ],
 )
