+++ src/cpp_accelerator/adapters/webrtc/protocol/data_channel_envelope.h
@@ -1,24 +1,22 @@
 #pragma once
 
-#include <cstdint>
-#include <string>
 #include <vector>
-
-#include <rtc/rtc.hpp>
 
 #include "proto/_virtual_imports/image_processor_service_proto/image_processor_service.pb.h"
 
 namespace jrb::adapters::webrtc::protocol {
 
+// Decodes a fully reassembled DataChannelRequest envelope.
+// Returns false if the bytes are neither a DataChannelRequest nor a legacy raw
+// ProcessImageRequest. On success, sets *is_keepalive iff the envelope was a
+// keepalive (process_request stays untouched), or fills *process_request with
+// the inner ProcessImageRequest.
 bool ParseDataChannelRequest(const std::vector<std::byte>& assembled,
                              cuda_learning::ProcessImageRequest* process_request,
                              bool* is_keepalive);
 
+// Carries api_version + trace_context from a request to its response.
 void CopyProcessMetadata(const cuda_learning::ProcessImageRequest& request,
                          cuda_learning::ProcessImageResponse* response);
-
-void SendFramed(rtc::DataChannel& dc, const std::string& payload, uint32_t message_id);
-
-int64_t CurrentUnixTimeMs();
 
 }  // namespace jrb::adapters::webrtc::protocol
